chore: move util traits to proper modules; remove inappropriate use of sqlx errors
