+++ src/datastore.rs
@@ -2,10 +2,9 @@ use crate::authorization::{Authorization, HeaderAuth, NoAuthorization};
 use crate::cache::PublicKeyState;
 use crate::consts::*;
 use crate::dime::Dime;
-use crate::domain::DimeProperties;
+use crate::domain::{DimeProperties, VecUtil};
 use crate::pb::public_key_request::Impl::HeaderAuth as HeaderAuthEnumRequest;
 use crate::pb::{PublicKeyRequest, public_key::Key};
-use crate::proto_helpers::VecUtil;
 use crate::storage::StoragePath;
 use crate::types::{OsError, Result};
 use prost::Message;
