+++ tests/object.rs
@@ -7,12 +7,13 @@ use linked_hash_map::LinkedHashMap;
 use object_store::cache::Cache;
 use object_store::config::Config;
 use object_store::datastore::{self, PublicKey, replication_object_uuids};
+use object_store::domain::{StringUtil, VecUtil};
 use object_store::pb::chunk::Impl::{Data, End};
 use object_store::pb::chunk_bidi::Impl::{
     Chunk as ChunkEnum, MultiStreamHeader as MultiStreamHeaderEnum,
 };
 
-use object_store::proto_helpers::{AudienceUtil, ObjectResponseUtil, StringUtil, VecUtil};
+use object_store::proto::{AudienceUtil, ObjectResponseUtil};
 use object_store::{consts::*, pb::HashRequest};
 
 use sqlx::PgPool;
